fix QUINT_* and warnings
git-svn-id: file:///home/justin/tmp/tmp/svn/ambrosia/trunk@16 168f9720-d8c2-4a15-ae6b-8c4eb8e64380

+++ iris/xmpp-core/stream.cpp
@@ -103,7 +103,7 @@ static QDomElement changeNS(const QDomElement &e, const QString &oldns, const QS
 
 	// copy attributes
 	QDomNamedNodeMap al = e.attributes();
-	for(x = 0; x < al.count(); ++x) {
+	for(x = 0; x < (uint)al.count(); ++x) {
 		QDomAttr a = al.item(x).cloneNode().toAttr();
 		i.setAttributeNodeNS(a);
 	}
